+++ app/components/bali/side_menu/preview.rb
@@ -5,11 +5,14 @@ module Bali
       # -------------------
       # Default menu with a section name and an item
       # @param title text
-      # @param name text
-      def default(title: 'Section Title', name: 'Item')
-        render(SideMenu::Component.new) do |c|
+      def default(title: 'Section Title')
+        render(SideMenu::Component.new(current_path: '/dashboard')) do |c|
           c.list(title: title) do |list|
-            list.item(name: name, href: '#')
+            list.item(name: 'Dashboard', href: '/dashboard')
+            list.item(name: 'Dashboard 2', href: '/t/dashboard')
+            list.item(name: 'Inventory', href: '/inventory')
+            list.item(name: 'Purchasing', href: '/purchasing')
+            list.item(name: 'Configuration', href: '/configuration')
           end
         end
       end
@@ -18,12 +21,12 @@ module Bali
       # -------------------
       # This will add an icon in the item specified
       # @param title text
-      # @param name text
-      # @param icon_name select [attachment, alert, paypal]
-      def with_icon(title: 'Section title', name: 'Item', icon_name: 'attachment')
-        render(SideMenu::Component.new) do |c|
+      def with_icon(title: 'Section title')
+        render(SideMenu::Component.new(current_path: '/attachment')) do |c|
           c.list(title: title) do |list|
-            list.item(name: name, href: '#', icon: icon_name)
+            list.item(name: 'Attachment', href: '/attachment', icon: 'attachment')
+            list.item(name: 'Alert', href: '/alert', icon: 'alert')
+            list.item(name: 'Paypal', href: '/paypal', icon: 'paypal')
           end
         end
       end
@@ -32,12 +35,13 @@ module Bali
       # -------------------
       # This will add a conditional to the item specified
       # @param title text
-      # @param name text
-      # @param authorized toggle
-      def authorized(title: 'Section title', name: 'Authorized Item', authorized: true)
-        render(SideMenu::Component.new) do |c|
+      def authorized(title: 'Section title')
+        render(SideMenu::Component.new(current_path: '/auth-1')) do |c|
           c.list(title: title) do |list|
-            list.item(name: name, authorized: authorized, href: '#')
+            list.item(name: 'Authorized 1', authorized: true, href: '/auth-1')
+            list.item(name: 'Authorized 2', authorized: true, href: '/auth-2')
+            list.item(name: 'Not authorized 1', authorized: false, href: '/not-auth-1')
+            list.item(name: 'Not authorized 2', authorized: false, href: '/not-auth-1')
           end
         end
       end
@@ -46,28 +50,30 @@ module Bali
       # -------------------
       # This will add subitems to an item
       # @param title text
-      # @param name text
-      # @param subitem_1 text
-      def with_sub_item(title: 'Section title', name: 'Item', subitem_1: 'Subitem 1')
-        render(SideMenu::Component.new) do |c|
+      def with_sub_item(title: 'Section title')
+        render(SideMenu::Component.new(current_path: '/parent-item')) do |c|
           c.list(title: title) do |list|
-            list.item(name: name, href: '#') do |item|
-              item.child_item(name: subitem_1, href: '#')
+            list.item(name: 'Parent Item', href: '/parent-item') do |item|
+              item.item(name: 'Child Item 1', href: '/child-item-1')
+              item.item(name: 'Child Item 2', href: '/child-item-2')
+              item.item(name: 'Child Item 3', href: '/child-item-3')
             end
           end
         end
       end
 
-      # SideMenu with item disabled
+      # SideMenu with Active Child Item
       # -------------------
-      # This will render a Menu with a disable item
+      # Will display parent as active when a child item is active
       # @param title text
-      # @param name text
-      # @param disabled toggle
-      def disabled(title: 'Section title', name: 'Item', disabled: true)
-        render(SideMenu::Component.new) do |c|
+      def with_active_sub_item(title: 'Section title')
+        render(SideMenu::Component.new(current_path: '/child-item-1')) do |c|
           c.list(title: title) do |list|
-            list.item(name: name, href: '#', disabled: disabled)
+            list.item(name: 'Parent Item', href: '/parent-item') do |item|
+              item.item(name: 'Child Item 1', href: '/child-item-1')
+              item.item(name: 'Child Item 2', href: '/child-item-2')
+              item.item(name: 'Child Item 3', href: '/child-item-3')
+            end
           end
         end
       end